requirements.txt
fastapi
uvicorn
pinecone
pydantic
python-dotenv
httpx
numpy
sentence-transformers
langgraph
langchain
mcp
kubernetes
PyGithub
requests
PyGithub
ragas
sentence-transformers
